telegram-site-chat/requirements.txt
aiogram==3.4.1
flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
flask-cors==4.0.0
pydantic-settings==2.1.0

